Renamed _logging to indicate privacy
