+++ setup.cfg
@@ -28,7 +28,7 @@ files =
     uvicorn/middleware/wsgi.py,
     tests/middleware/test_wsgi.py,
     uvicorn/supervisors/watchgodreload.py,
-    uvicorn/log.py,
+    uvicorn/_logging.py,
     uvicorn/middleware/asgi2.py,
     uvicorn/server.py,
     uvicorn/__init__.py,
