+++ uvicorn/protocols/http/httptools_impl.py
@@ -7,7 +7,7 @@ from collections import deque
 
 import httptools
 
-from uvicorn.log import TRACE_LOG_LEVEL
+from uvicorn._logging import TRACE_LOG_LEVEL
 from uvicorn.protocols.http.flow_control import (
     CLOSE_HEADER,
     HIGH_WATER_LIMIT,
